chore: cleanup dupes post-file move
