+++ src/utils/index.ts
@@ -15,13 +15,5 @@
  *
  */
 
-export { Git } from '../vcs/git';
-export {
-  GitCommandExecutor,
-  ensureQuoted,
-  isExecOutputSuccess,
-  sanitizeInput,
-  execCommand
-} from '../vcs/common';
 export { BaseHttpClient } from './base-http-client';
 export { Guards, isError, isTrue, isString } from './guards';
